feat: add general bottom navigation bar and general app bar
Created two seperate folders, AppBar and BottomBar, that contains the
the files related to the appbar and bottombar such as the actions
(buttons present in the bar) and the general widget itself. The general
widget only needs to be invoked as a widget with the arguments
containing its' dimensions if applicable and the buttons that would be
in the widget. The buttons are passed as an widget which are created and
stored in the general folders created, allowing for flexible navigation
bars and app bars.