feat: add public speaking routing and placeholders
Added widgets that would allow for the insertion of data relevant to the documented variables.
The Progression elements are still vague. In addition, another state is added in the Voquadro Controller
which is the Feedback State that allows a smooth control of pages to be shown using IndexedWidgets and switch statements.